Added a few more comments.
Explained what I mean by Rmm as a helped around dealing with the
values and locations pointed to by the ModR/M byte.

Also removed a switch that doesn't appear that i'll be needed since
we now look at the Op for any prefix on it,  Unless we decide to go
past 8086 at some point, not likely to be needed.