Move the modrm byte to the instruction
This way the length is correct as it knows what's been fetched
from memory

Pass the instruction to various methods so it's used consistently.
Ideally we can stop passing CPU to so many places as well

Clean up logging to make it cleaner, display flags like dosbox since
those look easy to parse

Use segment overrides when the prefix is present.